[FIR] Expand typealiases in supertypes
`FirSupertypesResolution` is backed by:
`FirPsiDiagnosticTestGenerated.testStatusResolveForTypealiasAsSuperClass`
(using `expandedTypeRef` of FIR makes
`NOTHING_TO_OVERRIDE` appear, because
there were no supertype scopes in `A`, because we
haven't yet committed the expanded type of the typealias,
but we certainly do need to check it first, as it may be a
random typealias from somewhere else)
and
`FirOldFrontendMPPDiagnosticsWithPsiTestGenerated.testExpectAnnotationAndInlineClassWithDefaultValue`
(otherwise we'll replace the implicit `Any` type ref with a regular
resolved type ref, which is incorrect).

Replacing `plus` with `assign` in `TypeUtils.kt`
was needed to make sure types obtained via
`typealias -> typealias -> class` would
contain the first typealias as the abbreviated
type. See:
`FirPsiOldFrontendDiagnosticsTestGenerated.testKt15734`
(`REDUNDANT_NULLABLE` disappears).