[FIR] Expand typealiases in a separate transformer
Consider:
`JvmIrInterpreterAfterFirPsi2IrTestGenerated#testAbstract1`.
In this test when transforming `object EmptyList`,
we need to expand `RandomAccess` (defined in
`compiler/testData/ir/interpreter/helpers/Collections.kt`),
but it doesn't yet have `expandedTypeRef`, and
neither we can call `getResolvedExpandedTypeRef(typeAlias)`,
nor `supertypeComputationSession.getResolvedSuperTypeRefsForOutOfSessionDeclaration()`.

But this is probably fine for `breakLoops()` as
it will manage to detect all loops sooner or
later when analyzing the last required file for
a given supertype loop.

We have to first wait until all the files are
processed.

^KT-65038