Stop ignoring `copy` during proto to stubs conversion
KT-11914

`StubBasedClassDeserialization` should deserialize `copy`, but before
this commit, it was generating it on the fly. Given that `copy` was
always the same it was okeish to generate it on the fly, but it becomes
a problem after KT-11914 is fixed because we can't generate correct
visibility for copy on the fly. We should get the real visibility from
the stubs. But to get it from the stubs, we should firstly start
serializing it to the stubs.

This commit:
1. Stops ignoring copy during proto (metadata?) to stubs conversion.
   CallableClsStubBuilder
2. Changes StubBasedClassDeserialization to stop generating `copy` on
   the fly. `copy` with correct visibility will be automatically
   deserialized from the stubs
3. `substitutedTypeFromLib` and `libraryDataClass_copy` tests now have
   the same output in IDE and standalone modes

There is no aparent reasons why we have been ignoring `copy`, so let's
just try to not ignore it.

Also see 47aa6392bfbd556a9e9c10a541e14416bcba1c34