[Analysis API] Provide proper implementation for jump targets
This commit changes semantics for 'hasMultipleJumpTargets'. Before,
it only returned 'false' if there are only 'return's or only loop jumps,
and if effective labels of those are the same.

Now, the behavior is corrected to match the KDoc – even if jump kinds
are different, 'hasMultipleJumpTargets' will be 'false' if both jumps
point to the same code place.