IJPL-938 [kotlin] Deprecate `createSmartPointer` util on `kotlin` side
This function is still needed in `kotlin` repo in some places,
but until the platform is updated, it cannot be thrown away.

There are only a handful of usages of that function in `kotlin`, so
it's OK to pessimize it in completion; but for `intellij` we need it
to be clearly marked as deprecated, so that it is not confused with
the platform one.

Making this function `internal` would have required copying it to
multiple modules, since it is used in at least four of them.
That seems like a worse option compared to a deprecation
with a clear message.